Fix integer mis-match warning

+++ RestrictEvents/SoftwareUpdate.cpp
@@ -173,7 +173,7 @@ static mach_vm_address_t org_sysctl_f16c;
 static int my_sysctl_f16c(__unused struct sysctl_oid *oidp, void *arg1, int arg2, struct sysctl_req *req) {
 	// Strip F16C bit from arg1
 	// Ref: https://github.com/apple-oss-distributions/xnu/blob/xnu-8020.101.4/bsd/kern/kern_mib.c#L935-L946
-	int mask = (uint64_t) (uintptr_t) arg1 & ~kHasF16C;
+	int mask = (int)((uint64_t)(uintptr_t)arg1 & ~(uint64_t)kHasF16C);
 
 	// Convert back
 	arg1 = (void *) (uintptr_t) mask;
